TIKA-4770: move DataURIScheme utils to tika-parser-datauri-commons
Per review: extraction wants commons-codec's lenient base64 (salvage bytes
from noisy data) rather than the JDK decoder's strictness, and tika-core
has no commons-codec. So the DataURIScheme classes move out of tika-core
into a new tika-parser-datauri-commons module (same pattern as
digest/jdbc/mail/xmp/zip-commons), package org.apache.tika.parser.datauri,
with the original commons-codec decode restored verbatim. The html and
text modules depend on it (replacing their direct commons-codec deps,
which are otherwise unused); listed in tika-bom.
