+++ tika-parsers/tika-parsers-standard/tika-parsers-standard-modules/tika-parser-html-module/pom.xml
@@ -45,8 +45,9 @@
     </dependency>
 
     <dependency>
-      <groupId>commons-codec</groupId>
-      <artifactId>commons-codec</artifactId>
+      <groupId>org.apache.tika</groupId>
+      <artifactId>tika-parser-datauri-commons</artifactId>
+      <version>${project.version}</version>
     </dependency>
 
     <!-- test scope -->
